% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cdfsum.copula.gauss}
\alias{cdfsum.copula.gauss}
\title{CDF of portfolio}
\usage{
cdfsum.copula.gauss(quantile, mu1, mu2, sigma1, sigma2, rho, nsteps)
}
\arguments{
\item{quantile}{portfolio P/L qunatile, number}

\item{mu1}{mean of first marginal P/L distribution, number}

\item{mu2}{mean of second marginal P/L distribution, number}

\item{sigma1}{standard deviation of first marginal P/L distribution, number, greater than zero}

\item{sigma2}{standard deviation of second marginal P/L distribution, number, greater than zero}

\item{rho}{measure of correlation between two assets, number, must be in $[0;1)$}

\item{nsteps}{the number of steps used in the copula approximation.
Increase in the number of steps increases the accuracy, but also increases the computation time.}
}
\description{
Derives P(X+Y<quantile) using Gaussian copula, normal margins. If X and Y are position P/Ls,
 then Value-at-Risk is equal to minus quantile.
}
\details{

}

