% Generated by roxygen2 (4.0.2): do not edit by hand
\name{copula.gauss}
\alias{copula.gauss}
\title{Gauss copula}
\usage{
copula.gauss(u, v, rho, nsteps)
}
\arguments{
\item{u}{value of frist marginal for random variable X: $P(X<x)=u$, number, $u in (0;1)$}

\item{v}{value of second marginal for random variable Y: $P(Y<v)=u$, number, $v in (0;1)$}

\item{rho}{correlation between X and Y, $rho in [0;1)$}

\item{nsteps}{number of steps used for each x and y}
}
\value{
value of copula in point $(u,v)$, number
}
\description{
Derives value of gaussian copula
}
\details{

}

